📦️ Move source code to src/cyclops/
