+++ src/cyclops/object_reader.py
@@ -5,8 +5,8 @@ Handles serialisation of objects.
 
 (c) Copyright UKAEA 2023.
 """
-import pickle
 import os
+import pickle
 
 
 class PickleManager:
