+++ src/cyclops/plotting.py
@@ -5,11 +5,11 @@ Handles plotting functions.
 
 (c) Copyright UKAEA 2023.
 """
+import os
+import numpy as np
 import matplotlib.pyplot as plt
 import matplotlib.cm as cm
 import scienceplots  # noqa F401, required for plt.style.use("science")
-import numpy as np
-import os
 
 
 class PlotManager:
