+++ src/cyclops/sensor_suite.py
@@ -5,9 +5,10 @@ Handles the various sensors employed by an experiment.
 
 (c) Copyright UKAEA 2023.
 """
-from sensors import Sensor
-from fields import Field
 import numpy as np
+
+from cyclops.sensors import Sensor
+from cyclops.fields import Field
 
 
 class SensorSuite:
